requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2
django-appconf==1.0.3
django-filebrowser==3.12.1
django-grappelli==2.13.3
django-imagekit==4.0.2
idna==2.8
pilkit==2.0
Pillow==7.0.0
pytz==2019.3
requests==2.22.0
six==1.14.0
sqlparse==0.3.0
urllib3==1.25.8
